backend/requirements.txt
Flask
flask-cors
requests  # For calling external APIs like Hugging Face or Gemini Pro
python-dotenv==0.19.0
google-generativeai==0.3.0
werkzeug==2.0.1
gunicorn==20.1.0